move api types to api-types crate
