+++ rust/api-types/Cargo.toml
@@ -7,7 +7,9 @@ license = "Apache-2.0"
 repository = "https://github.com/chroma-core/chroma"
 
 [dependencies]
+chroma-error.workspace = true
 serde.workspace = true
+thiserror.workspace = true
 utoipa = { version = "5.0.0", features = ["macros", "axum_extras", "debug", "uuid"], optional = true }
 
 [features]
