+++ rust/api-types/src/lib.rs
@@ -1,9 +1,11 @@
 mod error;
 mod fork;
+mod get;
 mod heartbeat;
 mod user_identity;
 
 pub use error::*;
 pub use fork::*;
+pub use get::*;
 pub use heartbeat::*;
 pub use user_identity::*;
